requirements.txt
openai-whisper>=20231117
moviepy>=1.0.3
gTTS>=2.3.2
transformers>=4.36.0
torch>=2.1.0
mutagen>=1.47.0
sentencepiece>=0.1.99
numpy>=1.24.0
tqdm>=4.66.0 